fix stacking of modals and sidebars
