+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@nordcloud/gnui",
   "description": "Nordcloud Design System - a collection of reusable React components used in Nordcloud's SaaS products",
-  "version": "7.9.0",
+  "version": "8.0.0",
   "license": "MIT",
   "repository": {
     "type": "git",
