+++ src/components/modal/index.tsx
@@ -1,2 +1,9 @@
 export * from "./Modal";
 export * from "./ModalConfirm";
+export {
+  ModalContent,
+  ModalHeader,
+  ModalActions,
+  ModalBox,
+  Background as ModalBackground,
+} from "./styles";
